requirement.txt
aiosmtplib==1.1.7
aiosqlite==0.17.0
anyio==3.6.2
bcrypt==4.0.1
blinker==1.5
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.87.0
fastapi-mail==1.2.2
h11==0.14.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
pydantic==1.10.2
PyJWT==2.6.0
pypika-tortoise==0.1.6
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
tortoise-orm==0.19.2
typing_extensions==4.4.0
uvicorn==0.20.0
